requirements.txt
utils
python-dotenv
yagmail